+++ NEWS.md
@@ -1,11 +1,13 @@
 The current version of the `MSEtool` package is available for download from [CRAN](https://CRAN.R-project.org/package=MSEtool).
 
-## MSEtool 3.0.3.9994 Development Version
+## MSEtool 3.0.3.9995 Development Version
 
 ### Fixes
 - fix the plus-group calculations in `popdynOneTScpp`
+- fix issue with simulated index in `AddInd` in the projection years
 
 ### New additions
+- add option to include empirical weight-at-age for the catches with `cpars$Wt_age_C`
 - F_crash, F_med, and other reference points have been added to `Hist@Ref$ByYear`
 - can now optimize for depletion in terms of SB/SBMSY with `control$D <- 'SBMSY'`
 - option to simulate CAL data including removals (retained + discards) with `OM@cpars$control$CAL="removals"`.